add wip PCI capability parsing script
Eventually want to integrate this into pciid so that we can decode all
the major capabilities we care about, representing them as structured
data.

There are some bugs in here right now, namely the length of the
capability regions should be more dynamic. Instead of using
offset + (next offset - offset), we should get the actual length
based on the capability's specifications. Most of them have headers
indicating length, and we should respect those.

Signed-off-by: Steven Noonan <steven@uplinklabs.net>